split up custom page into server and client component
