+++ app/custom/page.tsx
@@ -1,63 +1,13 @@
-'use client';
+import { VideoConferenceClientImpl } from './VideoConferenceClientImpl';
 
-import { formatChatMessageLinks, LiveKitRoom, VideoConference } from '@livekit/components-react';
-import {
-  ExternalE2EEKeyProvider,
-  LogLevel,
-  Room,
-  RoomConnectOptions,
-  RoomOptions,
-  VideoCodec,
-  VideoPresets,
-} from 'livekit-client';
-import { useSearchParams } from 'next/navigation';
-import { useMemo } from 'react';
-import { decodePassphrase } from '../../lib/client-utils';
-import { DebugMode } from '../../lib/Debug';
+interface SearchParams {
+  liveKitUrl?: string;
+  token?: string;
+  codec?: string;
+}
 
-export default function CustomRoomConnection() {
-  const searchParams = useSearchParams();
-  const liveKitUrl = searchParams?.get('liveKitUrl');
-  const token = searchParams?.get('token');
-  const codec = searchParams?.get('codec');
-
-  const e2eePassphrase =
-    typeof window !== 'undefined' && decodePassphrase(window.location.hash.substring(1));
-  const worker =
-    typeof window !== 'undefined' &&
-    new Worker(new URL('livekit-client/e2ee-worker', import.meta.url));
-  const keyProvider = new ExternalE2EEKeyProvider();
-
-  const e2eeEnabled = !!(e2eePassphrase && worker);
-
-  const roomOptions = useMemo((): RoomOptions => {
-    return {
-      publishDefaults: {
-        videoSimulcastLayers: [VideoPresets.h540, VideoPresets.h216],
-        red: !e2eeEnabled,
-        videoCodec: codec as VideoCodec | undefined,
-      },
-      adaptiveStream: { pixelDensity: 'screen' },
-      dynacast: true,
-      e2ee: e2eeEnabled
-        ? {
-            keyProvider,
-            worker,
-          }
-        : undefined,
-    };
-  }, []);
-
-  const room = useMemo(() => new Room(roomOptions), []);
-  if (e2eeEnabled) {
-    keyProvider.setKey(e2eePassphrase);
-    room.setE2EEEnabled(true);
-  }
-  const connectOptions = useMemo((): RoomConnectOptions => {
-    return {
-      autoSubscribe: true,
-    };
-  }, []);
+export default function CustomRoomConnection(props: { searchParams: SearchParams }) {
+  const { liveKitUrl, token, codec } = props.searchParams;
 
   if (typeof liveKitUrl !== 'string') {
     return <h2>Missing LiveKit URL</h2>;
@@ -68,19 +18,7 @@ export default function CustomRoomConnection() {
 
   return (
     <main data-lk-theme="default">
-      {liveKitUrl && (
-        <LiveKitRoom
-          room={room}
-          token={token}
-          connectOptions={connectOptions}
-          serverUrl={liveKitUrl}
-          audio={true}
-          video={true}
-        >
-          <VideoConference chatMessageFormatter={formatChatMessageLinks} />
-          <DebugMode logLevel={LogLevel.debug} />
-        </LiveKitRoom>
-      )}
+      <VideoConferenceClientImpl liveKitUrl={liveKitUrl} token={token} codec={codec} />
     </main>
   );
 }
